@@ -1,6 +1,6 @@
 {
-  "name": "RDM App",
-  "version": "0.5.0",
+  "name": "rdm-app",
+  "version": "0.8.7",
   "author": "Fortis IT-Services",
   "homepage": "https://fortis-it.de/",
   "scripts": {
@@ -15,40 +15,43 @@
   "dependencies": {
     "@angular/common": "~8.1.2",
     "@angular/compiler": "~8.1.2",
-    "@angular/core": "^6.0.3",
+    "@angular/core": "~8.1.2",
     "@angular/forms": "~8.1.2",
     "@angular/platform-browser": "~8.1.2",
     "@angular/platform-browser-dynamic": "~8.1.2",
     "@angular/router": "~8.1.2",
-    "@ionic-native/barcode-scanner": "^5.16.0",
-    "@ionic-native/camera": "^5.17.0",
-    "@ionic-native/core": "^5.0.0",
-    "@ionic-native/file": "^5.17.0",
-    "@ionic-native/file-path": "^5.17.0",
-    "@ionic-native/ionic-webview": "^5.17.0",
-    "@ionic-native/network": "^5.16.0",
-    "@ionic-native/splash-screen": "^5.0.0",
-    "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^4.7.1",
+    "@ionic-native/barcode-scanner": "^5.19.1",
+    "@ionic-native/camera": "^5.19.1",
+    "@ionic-native/core": "^5.19.1",
+    "@ionic-native/file": "^5.19.1",
+    "@ionic-native/file-path": "^5.19.1",
+    "@ionic-native/ionic-webview": "^5.19.1",
+    "@ionic-native/network": "^5.19.1",
+    "@ionic-native/photo-viewer": "^5.19.1",
+    "@ionic-native/splash-screen": "^5.19.1",
+    "@ionic-native/status-bar": "^5.19.1",
+    "@ionic/angular": "^4.11.7",
     "@ionic/storage": "^2.2.0",
+    "com-sarriaroman-photoviewer": "^1.2.4",
     "cordova": "^9.0.0",
     "cordova-android": "^8.1.0",
-    "cordova-ios": "5.1.1",
+    "cordova-ios": "^5.0.1",
     "cordova-plugin-camera": "^4.1.0",
     "cordova-plugin-file": "^6.0.2",
     "cordova-plugin-network-information": "^2.0.2",
     "cordova-sqlite-storage": "^3.4.0",
     "core-js": "^2.5.4",
-    "ionic": "^5.4.9",
+    "hammerjs": "^2.0.8",
+    "ionic": "^5.4.15",
     "ngx-qrcode2": "0.0.9",
     "phonegap-plugin-barcodescanner": "^8.1.0",
-    "rxjs": "~6.5.1",
+    "rxjs": "^6.5.4",
     "tslib": "^1.9.0",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/architect": "~0.801.2",
-    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular-devkit/build-angular": "^0.803.23",
     "@angular-devkit/core": "~8.1.2",
     "@angular-devkit/schematics": "~8.1.2",
     "@angular/cli": "~8.1.2",
@@ -59,7 +62,7 @@
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
+    "codelyzer": "^5.2.1",
     "cordova-plugin-device": "^2.0.2",
     "cordova-plugin-ionic-keyboard": "^2.2.0",
     "cordova-plugin-ionic-webview": "^4.1.3",
@@ -72,7 +75,7 @@
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "karma-jasmine-html-reporter": "^1.5.1",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
@@ -85,7 +88,9 @@
       "cordova-plugin-statusbar": {},
       "cordova-plugin-device": {},
       "cordova-plugin-splashscreen": {},
-      "cordova-plugin-ionic-webview": {},
+      "cordova-plugin-ionic-webview": {
+        "ANDROID_SUPPORT_ANNOTATIONS_VERSION": "27.+"
+      },
       "cordova-plugin-ionic-keyboard": {},
       "cordova-sqlite-storage": {},
       "cordova-plugin-network-information": {},
@@ -95,7 +100,8 @@
       "cordova-plugin-camera": {
         "ANDROID_SUPPORT_V4_VERSION": "27.+"
       },
-      "cordova-plugin-file": {}
+      "cordova-plugin-file": {},
+      "com-sarriaroman-photoviewer": {}
     },
     "platforms": [
       "ios",
