@@ -1,15 +1,16 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable, from, Subscriber } from 'rxjs';
+import { Observable, from } from 'rxjs';
 import { NetworkService, ConnectionStatus } from '../network/network.service';
-import { map, tap, catchError, filter, count } from 'rxjs/operators';
+import { map, tap, catchError } from 'rxjs/operators';
 import { OfflineService } from '../offline/offline.service';
 import { Storage } from '@ionic/storage';
 import { Chip } from '../../pages/parts/Chip';
 import { PartModel } from 'src/app/models/part/partmodel';
+import { API_IP } from './../../../environments/environment';
 
-const API_STORAGE_KEY = 'specialkey';
-const PART_URL = '../../../assets/data.json';
+const PART_URL = API_IP + 'parts/byProject/';
+const UPDATE_PART_URL = API_IP + 'parts';
 
 @Injectable({
   providedIn: 'root'
@@ -17,33 +18,82 @@ const PART_URL = '../../../assets/data.json';
 export class PartService {
 
   public items: PartModel[] = [];
+  public projectid: string;
 
   constructor(private http: HttpClient, private networkService: NetworkService, private storage: Storage, private offlineManager: OfflineService) { }
 
-  public getParts(forceRefresh: boolean = false, partId): Observable<any> {
-    if (this.networkService.getCurrentNetworkStatus() == ConnectionStatus.Offline || !forceRefresh) {
-      return from(this.getLocalData('parts'));
-    } else {
-      return this.http.get(`${PART_URL}` /*append partId*/).pipe(
+  public getParts(projectId): Observable<any> {
+    this.projectid = projectId;
+    if (this.networkService.getCurrentNetworkStatus() == ConnectionStatus.Offline) {
+      return from(this.getLocalData('parts'+this.projectid));
+    } 
+    else {
+      return this.http.get(`${PART_URL + projectId}`).pipe(
         map(res => res['parts']),
         tap(res => {
-          console.log('returns real live API data');
-          this.setLocalData('parts', res);
+          console.log('returns real live API data', PART_URL + projectId);
+          this.setLocalData('parts'+projectId, res);
           this.items = res;
         })
       );
     }
   }
 
+  public getOfflineParts() {
+    if (1 == 1) {
+      return this.getLocalData('parts'+this.projectid);
+    }
+  }
+
+  public setParts(partId, partItem) {
+    console.log('sets partdetail');
+    this.setLocalData('parts', partItem);
+  }
+
+  public createPart(data) {
+    let url = `${UPDATE_PART_URL}`;
+    this.items = [...this.items, data];
+    if (this.networkService.getCurrentNetworkStatus() == ConnectionStatus.Offline) {
+      this.setLocalData('parts'+this.projectid, this.items); //something went wrong here
+      return from(this.offlineManager.storeRequest(url, 'POST', data));
+    }
+    else {
+      this.http.post(url, data).subscribe(response => {
+          console.log(response);
+        },
+        error => {
+          alert(error);
+          console.log(error);
+      });
+      return this.http.post(url, data).pipe(catchError(err => {
+          this.offlineManager.storeRequest(url, 'POST', data);
+          throw new Error(err);
+        })
+      );
+    }
+  }
+
+  getDimensionsByFind(id) {
+    return this.items.find(x => x.counterId === id);
+  }
+
   //TODO
   public updatePart(data, partId): Observable<any> {
-    let url = `${PART_URL}`; ///parts/${partId}
-    console.log(url);
-
+    let url = `${UPDATE_PART_URL}`;
     if (this.networkService.getCurrentNetworkStatus() == ConnectionStatus.Offline) {
+      this.items[this.getDimensionsByFind(data.counterId).counterId -1] = data;
+      this.setLocalData('parts'+this.projectid, this.items); //something went wrong here
       return from(this.offlineManager.storeRequest(url, 'PUT', data));
     } 
     else {
+      this.http.put(url, data).subscribe(
+        response => {
+          console.log(response);
+        },
+        error => {
+          alert(error);
+          console.log(error);
+        });
       return this.http.put(url, data).pipe(catchError(err => {
           this.offlineManager.storeRequest(url, 'PUT', data);
           throw new Error(err);
@@ -52,69 +102,108 @@ export class PartService {
     }
   }
 
+
+  public deletePart(data): Observable<any>  {
+    let url = `${UPDATE_PART_URL + "/" + data.id}`;
+    if (this.networkService.getCurrentNetworkStatus() == ConnectionStatus.Offline) {
+      //this.removeLocalData(); //something went wrong here
+      let filtered = this.items.filter(x => {
+        return x != data;
+      });
+      console.log("old List", this.items);
+      console.log("new List", filtered);
+      this.items = filtered;
+      this.setLocalData('parts'+this.projectid, this.items);
+      return from(this.offlineManager.storeRequest(url, 'DELETE', data)); //todo Check if this works?
+    }
+    else {
+      this.http.delete(url).subscribe(
+          response => {
+            console.log(response);
+          },
+          error => {
+            alert(error);
+            console.log(error);
+          });
+      return this.http.delete(url).pipe(catchError(err => {
+            this.offlineManager.storeRequest(url, 'DELETE', data);
+            throw new Error(err);
+          })
+      );
+    }
+
+  }
+
   public filterItems(chips: Chip[]) :PartModel[] {
-    return this.items.filter(item => this.filterObj(chips, item));
+    if (chips.length > 0) 
+      return this.items.filter(item => this.filterObj(chips, item));
+    return this.items;
   }
 
   filterObj(chips: Chip[], item: PartModel): boolean {
     let ret = true;
     for (let chip of chips) {
+      let chipMatched = false;
       for (let term of chip.FilterTerm) {
         switch(chip.FilterObj) { 
           case "Ident-Nr": { 
-            if (item.id.toLowerCase().indexOf(term.toLowerCase()) > -1) {
-              ret = false;
+            if (item.counterId.toString().toLowerCase().includes(term.toLowerCase())) {
+              chipMatched = true;
             }
             break;
           } 
           case "P/N": { 
-            if(item.postModPN.toLowerCase().indexOf(term.toLowerCase()) > -1 == false) {
-              ret = false;
+            if (item.postModPN.toString().toLowerCase().includes(term.toLowerCase())) {
+              chipMatched = true;
             }
             break;
           } 
           case "Category": { 
-            if (item.componentType.toLowerCase().indexOf(term.toLowerCase()) > -1 == false) {
-              ret = false;
+            if (item.category.toString().toLowerCase().includes(term.toLowerCase())) {
+              chipMatched = true;
             }
             break;
           } 
-          case "componentType": {
-              if (item.componentType.toLowerCase().indexOf(term.toLowerCase()) > -1 == false) {
-                ret = false;
+          case "ComponentType": {
+              if (item.componentType.toString().toLowerCase().includes(term.toLowerCase())) {
+                chipMatched = true;
             }
             break;
           } 
           case "Status": { 
-            if (item.statusEdit.toLowerCase().indexOf(term.toLowerCase()) > -1 == false) {
-              ret = false;
+            if (item.rackLocation && item.rackNo && item.preModWeight && item.preModWeight != "N/A" && item.rackLocation != "N/A" && item.rackNo != "N/A") {
+              if (term === "Done") {
+                chipMatched = true;
+              }
+            }
+            else {
+              if (term === "ToDo") {
+                chipMatched = true;
+              }
             }
             break;
           } 
           case "Rack-Nr": { 
-            if (item.rackNo.toLowerCase().indexOf(term.toLowerCase()) > -1 == false) {
-              ret = false;
+            if (item.rackNo.toString().toLowerCase().includes(term.toLowerCase())) {
+              chipMatched = true;
             }
             break;
           } 
           case "Position": { 
-            if (item.postModPosition.toLowerCase().indexOf(term.toLowerCase()) > -1 == false) {
-              ret = false;
+            if (item.postModPosition.toString().toLowerCase().includes(term.toLowerCase())) {
+              chipMatched = true;
             }
             break;
-            
           } 
           case "InstallationRoom": { 
-            if (item.installZoneRoom.toLowerCase().indexOf(term.toLowerCase()) > -1 == false) {
-              ret = false;
+            if (item.installZoneRoom.toString().toLowerCase().includes(term.toLowerCase())) {
+              chipMatched = true;
             }
             break;
           } 
         }
       }
-      if (!ret) {
-        return ret;
-      }
+      ret = ret && chipMatched;
     }
     return ret;
   }
@@ -127,12 +216,19 @@ export class PartService {
 
   //Save result of API requests
   private setLocalData(key, data) {
-    this.storage.set(`${API_STORAGE_KEY}-${key}`, data);
+    this.storage.set(`${key}`, data);
   }
 
   //Get cached API result
   private getLocalData(key) {
     console.log("return local data");
-    return this.storage.get(`${API_STORAGE_KEY}-${key}`);
+    return this.storage.get(`${key}`);
+  }
+
+  //delete
+  private removeLocalData(){
+    this.storage.remove('parts').then(()=>{
+      console.log('part is removed');
+    });
   }
 }
