@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
-import { AuthService } from './services/auth/auth.service';
 
 const routes: Routes = [
   { path: '', redirectTo: 'login', pathMatch: 'full' },
@@ -8,6 +7,11 @@ const routes: Routes = [
   { path: 'projects', loadChildren: './pages/projects/projects.module#ProjectsPageModule' },
   { path: 'parts/:id', loadChildren: './pages/parts/parts.module#PartsPageModule' },
   { path: 'part-detail/:id', loadChildren: './pages/part-detail/part-detail.module#PartDetailPageModule' },
+  { path: 'part-detail/:id/:new', loadChildren: './pages/part-detail/part-detail.module#PartDetailPageModule' },
+  { path: 'popover', loadChildren: './component/popover/popover.module#PopoverPageModule' },
+  { path: 'help', loadChildren: './pages/help/help.module#HelpPageModule' },
+
+
 ];
 
 @NgModule({
