@@ -1,14 +1,16 @@
 import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
-import { ActionSheetController, Platform, ToastController } from '@ionic/angular';
+import { ActionSheetController, Platform, AlertController } from '@ionic/angular';
 import { Camera, CameraOptions, PictureSourceType } from '@ionic-native/camera/ngx';
 import { FilePath } from '@ionic-native/file-path/ngx';
-import { File, FileEntry } from '@ionic-native/file/ngx';
+import { File } from '@ionic-native/file/ngx';
 import { WebView } from '@ionic-native/ionic-webview/ngx';
 import { Storage } from '@ionic/storage';
 import { PartDetailPage } from 'src/app/pages/part-detail/part-detail.page';
 import { ToastService } from 'src/app/services/toast/toast.service';
-
-const STORAGE_KEY = 'my_images';
+import { PartService } from 'src/app/services/part/part.service';
+import { ProjectService } from 'src/app/services/project/project.service';
+import { ImageService } from 'src/app/services/image/image.service';
+import { PhotoViewer } from '@ionic-native/photo-viewer/ngx';
 
 @Component({
   selector: 'app-photo',
@@ -18,16 +20,39 @@ const STORAGE_KEY = 'my_images';
 export class PhotoComponent implements OnInit {
 
   images = [];
+  imagePath = "";
+  projectId: string;
 
-  constructor(private partDetail: PartDetailPage,private actionSheetController: ActionSheetController, private camera: Camera, private plt: Platform, private filePath: FilePath, private file: File, 
-    private toastController: ToastService, private webview: WebView, private storage: Storage, private ref: ChangeDetectorRef) { }
+  constructor(private alertCtrl: AlertController, private photoViewer: PhotoViewer, private imageService: ImageService, private projectService: ProjectService, private partDetail: PartDetailPage, private actionSheetController: ActionSheetController, private camera: Camera, private plt: Platform, private filePath: FilePath, private file: File, 
+    private toastController: ToastService, private webview: WebView, private storage: Storage, private ref: ChangeDetectorRef, private partService: PartService) { }
 
   ngOnInit() {
-    this.storage.get(STORAGE_KEY);
+    let partId = this.partDetail.counterId + 1;
+    this.projectId = this.projectService.getProjectId();
+    this.imagePath = "image/" + this.projectId + "/" + partId;
+    this.loadStoredImages();
+  }
+
+  openImage(url: string): void {
+    this.photoViewer.show(url);
+  }
+
+  loadStoredImages() {
+    this.storage.get(this.imagePath).then(images => {
+      if (images) {
+        let arr = JSON.parse(images);
+        this.images = [];
+        for (let img of arr) {
+          let filePath = this.file.dataDirectory + img;
+          let resPath = this.pathForImage(filePath);
+          this.images.push({ name: img, path: resPath, filePath: filePath });
+        }
+      }
+    });
   }
 
   async selectImage() {
-    this.partDetail.onSave(); //Speicher zwischen
+    this.partDetail.onSave(); // Speicher zwischen
     const actionSheet = await this.actionSheetController.create({
       header: "Select Image source",
       buttons: [{
@@ -54,7 +79,10 @@ export class PhotoComponent implements OnInit {
     var options: CameraOptions = {
       quality: 100,
       sourceType: sourceType,
-      saveToPhotoAlbum: false, //Nicht in Bibliothek
+      destinationType: this.camera.DestinationType.FILE_URI,
+      encodingType: this.camera.EncodingType.JPEG,
+      mediaType: this.camera.MediaType.PICTURE,
+      saveToPhotoAlbum: false, // Nicht in Bibliothek
       correctOrientation: true
     };
 
@@ -83,14 +111,14 @@ export class PhotoComponent implements OnInit {
   }
 
   updateStoredImages(name) {
-    this.storage.get(STORAGE_KEY).then(images => {
+    this.storage.get(this.imagePath).then(images => {
       let arr = JSON.parse(images);
       if (!arr) {
         let newImages = [name];
-        this.storage.set(STORAGE_KEY, JSON.stringify(newImages));
+        this.storage.set(this.imagePath, JSON.stringify(newImages));
       } else {
         arr.push(name);
-        this.storage.set(STORAGE_KEY, JSON.stringify(arr));
+        this.storage.set(this.imagePath, JSON.stringify(arr));
       }
 
       let filePath = this.file.dataDirectory + name;
@@ -101,44 +129,58 @@ export class PhotoComponent implements OnInit {
         path: resPath,
         filePath: filePath
       };
-      // ... BEDEUTET:
-      // const featured = ['Deep Dish', 'Pepperoni', 'Hawaiian'];
-      // const specialty = ['Meatzza', 'Spicy Mama', 'Margherita'];
-      // const pizzas = [...featured, 'veg pizza', ...specialty];
-      // console.log(pizzas); // 'Deep Dish', 'Pepperoni', 'Hawaiian', 'veg pizza', 'Meatzza', 'Spicy Mama', 'Margherita'
+
       this.images = [newEntry, ...this.images];
+      //this.partService.updatePart(this.images, this.partId);
+      this.imageService.uploadImage(newEntry, this.partDetail.getPartId(), this.imagePath);
       this.ref.detectChanges(); // trigger change detection cycle
     });
   }
 
-  deleteImage(imgEntry, position) {
-    this.images.splice(position, 1);
- 
-    this.storage.get(STORAGE_KEY).then(images => {
-      let arr = JSON.parse(images);
-      let filtered = arr.filter(name => name != imgEntry.name);
-      this.storage.set(STORAGE_KEY, JSON.stringify(filtered));
+  async deleteImage(imgEntry, position) {
+    const alert = await this.alertCtrl.create({
+      header: 'Confirm!',
+      message: 'Do you really want to delete this Image?',
+      buttons: [
+        {
+          text: 'Cancel',
+          handler: () => {
+            
+          }
+        }, {
+          text: 'Yes',
+          handler: () => {
+            this.images.splice(position, 1);
 
-      var correctPath = imgEntry.filePath.substr(0, imgEntry.filePath.lastIndexOf('/') + 1);
+            this.storage.get(this.imagePath).then(images => {
+              let arr = JSON.parse(images);
+              let filtered = arr.filter(name => name != imgEntry.name);
+              this.storage.set(this.imagePath, JSON.stringify(filtered));
 
-      this.file.removeFile(correctPath, imgEntry.name).then(res => {
-          this.toastController.displayToast('File removed.');
-      });
+              var correctPath = imgEntry.filePath.substr(0, imgEntry.filePath.lastIndexOf('/') + 1);
+
+              this.file.removeFile(correctPath, imgEntry.name).then(res => {
+                this.toastController.displayToast('File removed.');
+              });
+            });
+          }
+        }
+      ]
     });
+    await alert.present();
   }
 
   pathForImage(img) {
     if (img === null) {
       return '';
-    } 
-    else {
+    } else {
       let converted = this.webview.convertFileSrc(img);
       return converted;
     }
   }
 
   createFileName() {
-    return new Date().getTime() + ".jpg";
+    return new Date().getTime() + ".png";
   }
 
 }
